#191 Adjusted all tests
