+++ test/core_contract_test.exs
@@ -4,33 +4,8 @@ defmodule CoreContractTest do
   alias AeppSDK.{Account, Client, Contract, Utils.Keys}
 
   setup_all do
-    client =
-      Client.new(
-        %{
-          public: "ak_6A2vcm1Sz6aqJezkLCssUXcyZTX7X8D5UwbuS2fRJr9KkYpRU",
-          secret:
-            "a7a695f999b1872acb13d5b63a830a8ee060ba688a478a08c6e65dfad8a01cd70bb4ed7927f97b51e1bcb5e1340d12335b2a2b12c8bc5221d63c4bcb39d41e61"
-        },
-        "my_test",
-        "http://localhost:3013/v2",
-        "http://localhost:3113/v2"
-      )
-
-    source_code = "contract Identity =
-      datatype event = AddedNumberEvent(indexed int, string)
-
-      record state = { number : int }
-
-      entrypoint init(x : int) =
-        { number = x }
-
-      entrypoint get_number() =
-        state.number
-
-      stateful entrypoint add_to_number(x : int) =
-        Chain.event(AddedNumberEvent(x, \"Added a number\"))
-        state.number + x"
-    [client: client, source_code: source_code]
+    Code.require_file("test_utils.ex", "test/")
+    TestUtils.get_test_data()
   end
 
   @tag :travis_test
